+++ libs/data-mapper-v2/src/models/Schema.ts
@@ -1,0 +1,10 @@
+import type { SchemaType } from '@microsoft/logic-apps-shared';
+
+export interface FileWithVsCodePath extends File {
+  path?: string;
+}
+export interface SchemaFile {
+  name: string;
+  path: string;
+  type: SchemaType;
+}