+++ libs/designer-ui/src/lib/card/scopeCard/index.tsx
@@ -16,6 +16,7 @@ export interface ScopeCardProps extends CardProps {
 }
 
 export const ScopeCard: React.FC<ScopeCardProps> = ({
+  id,
   active = true,
   brandColor,
   collapsed,
@@ -33,7 +34,7 @@ export const ScopeCard: React.FC<ScopeCardProps> = ({
   onClick,
   onDeleteClick,
   handleCollapse,
-  selected,
+  selectionMode,
   contextMenuItems = [],
   runData,
   setFocus,
@@ -72,8 +73,6 @@ export const ScopeCard: React.FC<ScopeCardProps> = ({
           draggable={draggable}
           style={colorVars}
           onContextMenu={contextMenu.handle}
-          onKeyDown={keyboardInteraction.keyDown}
-          onKeyUp={keyboardInteraction.keyUp}
         >
           {isMonitoringView ? (
             <StatusPill
@@ -85,8 +84,15 @@ export const ScopeCard: React.FC<ScopeCardProps> = ({
             />
           ) : null}
           <div className="msla-scope-card-content">
-            <div className={css('msla-selection-box', 'white-outline', selected && 'selected')} />
-            <button className="msla-scope-card-title-button" ref={focusRef as any} onClick={handleClick}>
+            <div className={css('msla-selection-box', 'white-outline', selectionMode)} />
+            <button
+              id={`msla-node-${id}`}
+              className="msla-scope-card-title-button"
+              ref={focusRef as any}
+              onClick={handleClick}
+              onKeyDown={keyboardInteraction.keyDown}
+              onKeyUp={keyboardInteraction.keyUp}
+            >
               <div className="msla-scope-card-title-box">
                 <div className={css('gripper-section', draggable && 'draggable')}>{draggable ? <Gripper /> : null}</div>
                 <div className="panel-card-content-icon-section">{cardIcon}</div>